pyproject.toml
[tool.poetry]
name = "test_taak_QA"
version = "0.1.0"
description = "there is a  test  task solution"
authors = ["Arner Pan <sergalbox@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
selenium = "^4.3.0"
allure-pytest = "^2.9.45"
chromedriver-binary = "^104.0.5112"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
